+++ spec/views/essences/essence_picture_view_spec.rb
@@ -1,13 +1,22 @@
 require 'spec_helper'
 
 describe "essences/_essence_picture_view" do
-  let(:essence_picture) { stub_model(Alchemy::EssencePicture, picture: stub_model(Alchemy::Picture), caption: 'This is a cute cat') }
-  let(:content) { stub_model(Alchemy::Content, name: 'image', essence_type: 'EssencePicture', essence: essence_picture) }
+  let(:essence_picture) do
+    stub_model Alchemy::EssencePicture,
+      picture: stub_model(Alchemy::Picture, image_file_format: 'jpg'),
+      caption: 'This is a cute cat'
+  end
+
+  let(:content) do
+    stub_model Alchemy::Content,
+      name: 'image',
+      essence_type: 'EssencePicture',
+      essence: essence_picture
+  end
 
   before do
     ActionView::Base.send(:include, Alchemy::UrlHelper)
     ActionView::Base.send(:include, Alchemy::EssencesHelper)
-    allow(view).to receive(:configuration).and_return(:jpg)
   end
 
   context "with caption" do